requirements.txt
dnspython>=2.0.0
tqdm>=4.0.0
